BigDataCluster/requirements.txt
apache-airflow==2.7.2
apache-airflow-providers-apache-spark==4.1.1
hdfs==2.7.3
kafka-python==2.0.2
pyspark==3.5.0
